requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.5
django-filter==2.2.0
django-heroku==0.3.1
django-toolbelt==0.0.1
djangorestframework==3.11.0
gunicorn==20.0.4
httpie==2.1.0
idna==2.9
itypes==1.2.0
Jinja2==2.11.2
Markdown==3.2.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
psycopg2==2.8.5
Pygments==2.6.1
pytz==2019.3
requests==2.23.0
simplejson==3.17.0
sqlparse==0.3.1
static3==0.7.0
uritemplate==3.0.1
urllib3==1.25.9
whitenoise==5.0.1
